fix: SignalGeneratorの出力がfloatではない問題の修正

+++ sifigan/utils/features.py
@@ -116,6 +116,7 @@ class SignalGenerator:
         input_batch = signals[0]
         for signal in signals[1:]:
             input_batch = torch.cat([input_batch, signal], axis=1)
+        input_batch = input_batch.float()
 
         return input_batch
 
@@ -148,7 +149,9 @@ class SignalGenerator:
         """
         B, _, T = f0.size()
         vuv = interpolate((f0 > 0) * torch.ones_like(f0), T * self.hop_size)
-        radious = (interpolate(f0.to(torch.float64), T * self.hop_size) / self.sample_rate) % 1
+        radious = (
+            interpolate(f0.to(torch.float64), T * self.hop_size) / self.sample_rate
+        ) % 1
         sine = vuv * torch.sin(torch.cumsum(radious, dim=2) * 2 * np.pi) * self.sine_amp
         if self.noise_amp > 0:
             noise_amp = vuv * self.noise_amp + (1.0 - vuv) * self.noise_amp / 3.0
